timeOfLastClick to disable scroll nav effect when nav is clicked
Prevents the animation being triggered on page scroll, when the nav button is triggered for one second. This way when you click a nav link, the page scrolls and the animation only goes from A to C. Whereas previously the page would scroll down and show the animation from B, then jump over to C. This looks alot smoother as the scroll animation is disabled for one second whilst the page scrolls on click.